+++ src/auth/oauth/test/oauth.test.ts
@@ -24,33 +24,6 @@ beforeEach(() => {
   (Cookies as any).mockClear();
 });
 
-declare global {
-  namespace jest {
-    interface Matchers<R> {
-      toBeWithinSecondsOf(compareDate: number, seconds: number): R;
-    }
-  }
-}
-
-expect.extend({
-  toBeWithinSecondsOf(received: number, compareDate: number, seconds) {
-    const pass = received && compareDate && Math.abs(received - compareDate) <= seconds * 1000;
-    if (pass) {
-      return {
-        message: () =>
-          `expected ${received} not to be within ${seconds} seconds of ${compareDate}`,
-        pass: true,
-      };
-    } else {
-      return {
-        message: () =>
-          `expected ${received} to be within ${seconds} seconds of ${compareDate}`,
-        pass: false,
-      };
-    }
-  },
-});
-
 describe('beginAuth', () => {
   let cookies = {
     id: '',
@@ -580,6 +553,7 @@ describe('validateAuthCallback', () => {
     fetchMock.mockResponse(JSON.stringify(successResponse));
     const returnedSession = await ShopifyOAuth.validateAuthCallback(req, res, testCallbackQuery);
     expect(returnedSession.id).toEqual(cookies.id);
+    expect(returnedSession.id).toEqual(ShopifyOAuth.getOfflineSessionId(shop));
 
     const cookieSession = await Context.SESSION_STORAGE.loadSession(cookies.id);
     expect(cookieSession).not.toBeUndefined();